src/requirements.txt
anyio==4.3.0
asgiref==3.8.1
certifi==2024.2.2
Django==5.0.3
djangorestframework==3.15.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
psycopg==3.1.18
psycopg-binary==3.1.18
redis==5.0.3
sniffio==1.3.1
sqlparse==0.4.4
typing_extensions==4.10.0
